1. webRTC支持VP9
git-svn-id: http://10.22.1.27:8000/svn/rooten/MSMS/src/client/IOS/IM@2639 47c12b7e-2652-5545-b168-31c46e2fce3f

+++ IM/webrtc/RTCICECandidate+WRJSON.m
@@ -27,7 +27,7 @@ static NSString const *kRTCICECandidateSdpKey = @"candidate";
 - (NSData *)JSONData {
     NSDictionary *json = @{
                            kRTCICECandidateTypeKey : kRTCICECandidateTypeValue,
-                           kRTCICECandidateMLineIndexKey : [NSString stringWithFormat:@"%d", self.sdpMLineIndex],
+                           kRTCICECandidateMLineIndexKey : [NSString stringWithFormat:@"%ld", (long)self.sdpMLineIndex],
                            kRTCICECandidateMidKey : self.sdpMid,
                            kRTCICECandidateSdpKey : self.sdp
                            };
